requirements.txt
# ── Core Framework ────────────────────────────────────────────
Django==5.2.8
asgiref==3.10.0

# ── Django Extensions ─────────────────────────────────────────
django-allauth==65.13.1
django-import-export==4.3.14
django-nested-admin==4.1.6
django-tailwind==4.4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
dj-database-url==3.0.1

# ── API Layer ─────────────────────────────────────────────────
django-ninja==1.3.0
django-cors-headers==4.4.0

# ── Database ──────────────────────────────────────────────────
psycopg2-binary==2.9.11
mysqlclient==2.2.7

# ── Storage & Static Files ────────────────────────────────────
django-storages[azure]==1.14.6
azure-storage-blob==12.28.0
whitenoise==6.11.0
pillow==12.0.0

# ── Server ────────────────────────────────────────────────────
gunicorn==23.0.0

# ── Environment & Config ──────────────────────────────────────
python-dotenv==1.2.1
python-decouple==3.8
pytailwindcss==0.3.0

# ── Auth ──────────────────────────────────────────────────────
PyJWT==2.10.1

# ── Vocabulary / FSRS ─────────────────────────────────────────
fsrs==6.3.0
pykakasi==2.2.1
jaconv==0.4.1

# ── AI / External APIs ────────────────────────────────────────
google-generativeai==0.8.6
requests==2.32.3

# ── Utilities ─────────────────────────────────────────────────
beautifulsoup4==4.14.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
qrcode[pil]==7.4.2
PyYAML==6.0.2
